+++ src/Service/GamblingTecRNG.php
@@ -1,26 +1,32 @@
 <?php
+/**
+ * Copyright Sunseven NN,
+ * E-Commerce Park, Willemstad, Curacao
+ * To be used according to the prevailing terms and conditions
+ * as set out in the discourse with Sunseven NV.
+ */
 namespace Application\Service;
 
 use Zend\Math\Rand;
 
 class GamblingTecRNG
 {
-    public function getInteger(int $min, int $max)
+    public static function getInteger(int $min, int $max)
     {
         return Rand::getInteger($min, $max);
     }
 
-    public function getBoolean()
+    public static function getBoolean()
     {
         return Rand::getBoolean();
     }
 
-    public function getFloat()
+    public static function getFloat()
     {
         return Rand::getFloat();
     }
 
-    public function getString($min, $charlist = null)
+    public static function getString($min, $charlist = null)
     {
         return Rand::getString($min, $charlist);
     }