+++ CHANGELOG.md
@@ -6,10 +6,22 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 
 ## [Unreleased]
 
+## [0.1.1]
+### Fixed
+- General content spelling mistakes
+
+### Added
+- Fisher Yates card shuffle
+- More detailed content
+- Clubs, Diamonds, Hearts, Spades game
+- Scaled data method
+- General copyright data
+
 ## [0.1.0]
 ### Added
 - Initial release
 - This CHANGELOG file
 
-[Unreleased]: https://github.com/chateaux/gamblingtec-rng/compare/0.1.0...HEAD
+[Unreleased]: https://github.com/chateaux/gamblingtec-rng/compare/0.1.1...HEAD
+[0.1.1]: https://github.com/chateaux/gamblingtec-rng/compare/0.1.0...0.1.1
 [0.1.0]: https://github.com/chateaux/gamblingtec-rng/compare/039242477f4ba5f7ed6d6fc735e9349a913dbba7...0.1.0